re-design single property editor
