--- src/desktop/AwbStudio/ProjectConfiguration/PropertyEditors/ValueEditorControl.xaml
@@ -1,17 +1,0 @@
-﻿<UserControl x:Class="AwbStudio.ProjectConfiguration.PropertyEditors.ValueEditorControl"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:AwbStudio.ProjectConfiguration.PropertyEditors"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="auto" MinWidth="100"></ColumnDefinition>
-            <ColumnDefinition Width="*"></ColumnDefinition>
-        </Grid.ColumnDefinitions>
-        <Label Grid.Column="0" Content="{Binding PropertyTitle, RelativeSource={RelativeSource AncestorType=UserControl}}" HorizontalAlignment="Right" VerticalAlignment="Center"></Label>
-        <TextBox Grid.Column="1" Text="{Binding PropertyContent, Mode=TwoWay, RelativeSource={RelativeSource AncestorType=UserControl}}"></TextBox>
-    </Grid>
-</UserControl>
